saccoVoteAPI/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
Command==0.1.0
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.1
django-allauth==0.61.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
drf-nested-routers==0.93.5
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.1
